Separate test and model code
